Fixed all warnings on MSVC

+++ draw.h
@@ -8,8 +8,8 @@ index = MIN(index, fixedSize - 1);
 
 if (t < t_corona) {
   glUseProgram(shader_program_gfx_logo210.handle);
-  glUniform1f(shader_uniform_gfx_logo210_iTime, t - t_logo210);
-  glUniform2f(shader_uniform_gfx_logo210_iResolution, w, h);
+  glUniform1f(shader_uniform_gfx_logo210_iTime, (float)(t - t_logo210));
+  glUniform2f(shader_uniform_gfx_logo210_iResolution, (float)w, (float)h);
 #ifdef MIDI
   glUniform1f(shader_uniform_gfx_logo210_iFader0, fader0);
   glUniform1f(shader_uniform_gfx_logo210_iFader1, fader1);
@@ -22,8 +22,8 @@ if (t < t_corona) {
 #endif
 } else if (t < t_menger) {
   glUseProgram(shader_program_gfx_corona.handle);
-  glUniform1f(shader_uniform_gfx_corona_iTime, t - t_corona);
-  glUniform2f(shader_uniform_gfx_corona_iResolution, w, h);
+  glUniform1f(shader_uniform_gfx_corona_iTime, (float)(t - t_corona));
+  glUniform2f(shader_uniform_gfx_corona_iResolution, (float)w, (float)h);
 #ifdef MIDI
   glUniform1f(shader_uniform_gfx_corona_iFader0, fader0);
   glUniform1f(shader_uniform_gfx_corona_iFader1, fader1);
@@ -36,8 +36,8 @@ if (t < t_corona) {
 #endif
 } else if (t < t_blockmania) {
   glUseProgram(shader_program_gfx_menger.handle);
-  glUniform1f(shader_uniform_gfx_menger_iTime, t - t_menger);
-  glUniform2f(shader_uniform_gfx_menger_iResolution, w, h);
+  glUniform1f(shader_uniform_gfx_menger_iTime, (float)(t - t_menger));
+  glUniform2f(shader_uniform_gfx_menger_iResolution, (float)w, (float)h);
 
 #ifdef MIDI
   glUniform1f(shader_uniform_gfx_menger_iFader0, fader0);
@@ -50,12 +50,12 @@ if (t < t_corona) {
   glUniform1f(shader_uniform_gfx_menger_iFader7, fader7);
 #endif
 
-  glUniform1f(shader_uniform_gfx_menger_iScale, MAX(MIN(scale[index], 1.), 0.));
+  glUniform1f(shader_uniform_gfx_menger_iScale, MAX(MIN(scale[index], 1.f), 0.f));
   // glUniform1f(shader_uniform_gfx_menger_iNBeats, nBeats[index]);
 } else if (t < t_edgysphere) {
   glUseProgram(shader_program_gfx_blockmania.handle);
-  glUniform1f(shader_uniform_gfx_blockmania_iTime, t - t_blockmania);
-  glUniform2f(shader_uniform_gfx_blockmania_iResolution, w, h);
+  glUniform1f(shader_uniform_gfx_blockmania_iTime, (float)(t - t_blockmania));
+  glUniform2f(shader_uniform_gfx_blockmania_iResolution, (float)w, (float)h);
 #ifdef MIDI
   glUniform1f(shader_uniform_gfx_blockmania_iFader0, fader0);
   glUniform1f(shader_uniform_gfx_blockmania_iFader1, fader1);
@@ -67,12 +67,12 @@ if (t < t_corona) {
   glUniform1f(shader_uniform_gfx_blockmania_iFader7, fader7);
 #endif
 
-  glUniform1f(shader_uniform_gfx_blockmania_iScale, MAX(MIN(scale[index], 1.), 0.));
+  glUniform1f(shader_uniform_gfx_blockmania_iScale, MAX(MIN(scale[index], 1.f), 0.f));
   // glUniform1f(shader_uniform_gfx_blockmania_iNBeats, nBeats[index]);
 } else if (t < t_tentacles) {
   glUseProgram(shader_program_gfx_edgysphere.handle);
-  glUniform1f(shader_uniform_gfx_edgysphere_iTime, t - t_edgysphere);
-  glUniform2f(shader_uniform_gfx_edgysphere_iResolution, w, h);
+  glUniform1f(shader_uniform_gfx_edgysphere_iTime, (float)(t - t_edgysphere));
+  glUniform2f(shader_uniform_gfx_edgysphere_iResolution, (float)w, (float)h);
 #ifdef MIDI
   glUniform1f(shader_uniform_gfx_edgysphere_iFader0, fader0);
   glUniform1f(shader_uniform_gfx_edgysphere_iFader1, fader1);
@@ -84,12 +84,12 @@ if (t < t_corona) {
   glUniform1f(shader_uniform_gfx_edgysphere_iFader7, fader7);
 #endif
 
-  glUniform1f(shader_uniform_gfx_edgysphere_iScale, MAX(MIN(scale[index], 1.), 0.));
+  glUniform1f(shader_uniform_gfx_edgysphere_iScale, MAX(MIN(scale[index], 1.f), 0.f));
   // glUniform1f(shader_uniform_gfx_edgysphere_iNBeats, nBeats[index]);
 } else {
   glUseProgram(shader_program_gfx_tentacles.handle);
-  glUniform1f(shader_uniform_gfx_tentacles_iTime, t - t_tentacles);
-  glUniform2f(shader_uniform_gfx_tentacles_iResolution, w, h);
+  glUniform1f(shader_uniform_gfx_tentacles_iTime, (float)(t - t_tentacles));
+  glUniform2f(shader_uniform_gfx_tentacles_iResolution, (float)w, (float)h);
 #ifdef MIDI
   glUniform1f(shader_uniform_gfx_tentacles_iFader0, fader0);
   glUniform1f(shader_uniform_gfx_tentacles_iFader1, fader1);
@@ -101,7 +101,7 @@ if (t < t_corona) {
   glUniform1f(shader_uniform_gfx_tentacles_iFader7, fader7);
 #endif
 
-  glUniform1f(shader_uniform_gfx_tentacles_iScale, MAX(MIN(scale[index], 1.), 0.));
+  glUniform1f(shader_uniform_gfx_tentacles_iScale, MAX(MIN(scale[index], 1.f), 0.f));
   // glUniform1f(shader_uniform_gfx_tentacles_iNBeats, nBeats[index]);
 }
 #endif
